Initial commit
This checking represents a base level of functionality. 
The built tools were the WinDDK 3790.1830, NASM 2.12.01, and the Intel ASL compiler.
It was built as a debug target for x64 arch and tested in the UEFI shell of a system with native UEFI firmware. 

More validation is needed but this is a starting point.